+++ administrator/components/com_installer/View/Install/Html.php
@@ -1,0 +1,63 @@
+<?php
+/**
+ * @package     Joomla.Administrator
+ * @subpackage  com_installer
+ *
+ * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+namespace Joomla\Component\Installer\Administrator\View\Install;
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Component\ComponentHelper;
+use Joomla\CMS\Plugin\PluginHelper;
+use Joomla\CMS\Toolbar\ToolbarHelper;
+use Joomla\Component\Installer\Administrator\View\Installer\Html as InstallerViewDefault;
+
+/**
+ * Extension Manager Install View
+ *
+ * @since  1.5
+ */
+class Html extends InstallerViewDefault
+{
+	/**
+	 * Display the view
+	 *
+	 * @param   string  $tpl  Template
+	 *
+	 * @return  void
+	 *
+	 * @since   1.5
+	 */
+	public function display($tpl = null)
+	{
+		$paths        = new \stdClass;
+		$paths->first = '';
+
+		$this->paths  = &$paths;
+
+		$this->showJedAndWebInstaller = ComponentHelper::getParams('com_installer')->get('show_jed_info', 1);
+
+		PluginHelper::importPlugin('installer');
+
+		\JFactory::getApplication()->triggerEvent('onInstallerBeforeDisplay', array(&$this->showJedAndWebInstaller, $this));
+
+		parent::display($tpl);
+	}
+
+	/**
+	 * Add the page title and toolbar.
+	 *
+	 * @return  void
+	 *
+	 * @since   1.6
+	 */
+	protected function addToolbar()
+	{
+		parent::addToolbar();
+
+		ToolbarHelper::help('JHELP_EXTENSIONS_EXTENSION_MANAGER_INSTALL');
+	}
+}
