+++ administrator/components/com_installer/View/Update/Html.php
@@ -1,0 +1,86 @@
+<?php
+/**
+ * @package     Joomla.Administrator
+ * @subpackage  com_installer
+ *
+ * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+namespace Joomla\Component\Installer\Administrator\View\Update;
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Toolbar\ToolbarHelper;
+use Joomla\Component\Installer\Administrator\View\Installer\Html as InstallerViewDefault;
+
+/**
+ * Extension Manager Update View
+ *
+ * @since  1.6
+ */
+class Html extends InstallerViewDefault
+{
+	/**
+	 * List of update items.
+	 *
+	 * @var array
+	 */
+	protected $items;
+
+	/**
+	 * List pagination.
+	 *
+	 * @var \Joomla\CMS\Pagination\Pagination
+	 */
+	protected $pagination;
+
+	/**
+	 * Display the view.
+	 *
+	 * @param   string  $tpl  Template
+	 *
+	 * @return  void
+	 *
+	 * @since   1.6
+	 */
+	public function display($tpl = null)
+	{
+		// Get data from the model.
+		$this->items         = $this->get('Items');
+		$this->pagination    = $this->get('Pagination');
+		$this->filterForm    = $this->get('FilterForm');
+		$this->activeFilters = $this->get('ActiveFilters');
+
+		$paths        = new \stdClass;
+		$paths->first = '';
+
+		$this->paths = &$paths;
+
+		if (count($this->items) > 0)
+		{
+			\JFactory::getApplication()->enqueueMessage(\JText::_('COM_INSTALLER_MSG_WARNINGS_UPDATE_NOTICE'), 'notice');
+		}
+
+		parent::display($tpl);
+	}
+
+	/**
+	 * Add the page title and toolbar.
+	 *
+	 * @return  void
+	 *
+	 * @since   1.6
+	 */
+	protected function addToolbar()
+	{
+		ToolbarHelper::custom('update.update', 'upload', 'upload', 'COM_INSTALLER_TOOLBAR_UPDATE', true);
+		ToolbarHelper::custom('update.find', 'refresh', 'refresh', 'COM_INSTALLER_TOOLBAR_FIND_UPDATES', false);
+		ToolbarHelper::custom('update.purge', 'purge', 'purge', 'COM_INSTALLER_TOOLBAR_PURGE', false);
+		ToolbarHelper::divider();
+
+		\JHtmlSidebar::setAction('index.php?option=com_installer&view=manage');
+
+		parent::addToolbar();
+		ToolbarHelper::help('JHELP_EXTENSIONS_EXTENSION_MANAGER_UPDATE');
+	}
+}
