+++ administrator/components/com_installer/View/Updatesites/Html.php
@@ -1,0 +1,98 @@
+<?php
+/**
+ * @package     Joomla.Administrator
+ * @subpackage  com_installer
+ *
+ * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+namespace Joomla\Component\Installer\Administrator\View\Updatesites;
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Helper\ContentHelper;
+use Joomla\CMS\Toolbar\ToolbarHelper;
+use Joomla\Component\Installer\Administrator\View\Installer\Html as InstallerViewDefault;
+
+/**
+ * Extension Manager Update Sites View
+ *
+ * @package     Joomla.Administrator
+ * @subpackage  com_installer
+ * @since       3.4
+ */
+class Html extends InstallerViewDefault
+{
+	protected $items;
+
+	protected $pagination;
+
+	protected $form;
+
+	/**
+	 * Display the view
+	 *
+	 * @param   string  $tpl  Template
+	 *
+	 * @return  mixed|void
+	 *
+	 * @since   3.4
+	 *
+	 * @throws  \Exception on errors
+	 */
+	public function display($tpl = null)
+	{
+		// Get data from the model
+		$this->items         = $this->get('Items');
+		$this->pagination    = $this->get('Pagination');
+		$this->filterForm    = $this->get('FilterForm');
+		$this->activeFilters = $this->get('ActiveFilters');
+
+		// Check for errors.
+		if (count($errors = $this->get('Errors')))
+		{
+			throw new \JViewGenericdataexception(implode("\n", $errors), 500);
+		}
+
+		// Include the component HTML helpers.
+		\JHtml::addIncludePath(JPATH_COMPONENT . '/helpers/html');
+
+		// Display the view
+		parent::display($tpl);
+	}
+
+	/**
+	 * Add the page title and toolbar.
+	 *
+	 * @return  void
+	 *
+	 * @since   3.4
+	 */
+	protected function addToolbar()
+	{
+		$canDo = ContentHelper::getActions('com_installer');
+
+		if ($canDo->get('core.edit.state'))
+		{
+			ToolbarHelper::publish('updatesites.publish', 'JTOOLBAR_ENABLE', true);
+			ToolbarHelper::unpublish('updatesites.unpublish', 'JTOOLBAR_DISABLE', true);
+			ToolbarHelper::divider();
+		}
+
+		if ($canDo->get('core.delete'))
+		{
+			ToolbarHelper::deleteList('JGLOBAL_CONFIRM_DELETE', 'updatesites.delete', 'JTOOLBAR_DELETE');
+			ToolbarHelper::divider();
+		}
+
+		if ($canDo->get('core.admin') || $canDo->get('core.options'))
+		{
+			ToolbarHelper::custom('updatesites.rebuild', 'refresh.png', 'refresh_f2.png', 'JTOOLBAR_REBUILD', false);
+		}
+
+		\JHtmlSidebar::setAction('index.php?option=com_installer&view=updatesites');
+
+		parent::addToolbar();
+		ToolbarHelper::help('JHELP_EXTENSIONS_EXTENSION_MANAGER_UPDATESITES');
+	}
+}
