--- administrator/components/com_installer/controllers/discover.php
@@ -1,59 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_installer
- *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-
-/**
- * Discover Installation Controller
- *
- * @since  1.6
- */
-class InstallerControllerDiscover extends JControllerLegacy
-{
-	/**
-	 * Refreshes the cache of discovered extensions.
-	 *
-	 * @return  void
-	 *
-	 * @since   1.6
-	 */
-	public function refresh()
-	{
-		$model = $this->getModel('discover');
-		$model->discover();
-		$this->setRedirect(JRoute::_('index.php?option=com_installer&view=discover', false));
-	}
-
-	/**
-	 * Install a discovered extension.
-	 *
-	 * @return  void
-	 *
-	 * @since   1.6
-	 */
-	public function install()
-	{
-		$this->getModel('discover')->discover_install();
-		$this->setRedirect(JRoute::_('index.php?option=com_installer&view=discover', false));
-	}
-
-	/**
-	 * Clean out the discovered extension cache.
-	 *
-	 * @return  void
-	 *
-	 * @since   1.6
-	 */
-	public function purge()
-	{
-		$model = $this->getModel('discover');
-		$model->purge();
-		$this->setRedirect(JRoute::_('index.php?option=com_installer&view=discover', false), $model->_message);
-	}
-}
