+++ administrator/components/com_installer/views/installer/tmpl/default_ftp.php
