--- administrator/components/com_installer/views/updatesites/view.html.php
@@ -1,95 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_installer
- *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-
-JLoader::register('InstallerViewDefault', dirname(__DIR__) . '/default/view.php');
-
-/**
- * Extension Manager Update Sites View
- *
- * @package     Joomla.Administrator
- * @subpackage  com_installer
- * @since       3.4
- */
-class InstallerViewUpdatesites extends InstallerViewDefault
-{
-	protected $items;
-
-	protected $pagination;
-
-	protected $form;
-
-	/**
-	 * Display the view
-	 *
-	 * @param   string  $tpl  Template
-	 *
-	 * @return  mixed|void
-	 *
-	 * @since   3.4
-	 *
-	 * @throws  Exception on errors
-	 */
-	public function display($tpl = null)
-	{
-		// Get data from the model
-		$this->items         = $this->get('Items');
-		$this->pagination    = $this->get('Pagination');
-		$this->filterForm    = $this->get('FilterForm');
-		$this->activeFilters = $this->get('ActiveFilters');
-
-		// Check for errors.
-		if (count($errors = $this->get('Errors')))
-		{
-			throw new JViewGenericdataexception(implode("\n", $errors), 500);
-		}
-
-		// Include the component HTML helpers.
-		JHtml::addIncludePath(JPATH_COMPONENT . '/helpers/html');
-
-		// Display the view
-		parent::display($tpl);
-	}
-
-	/**
-	 * Add the page title and toolbar.
-	 *
-	 * @return  void
-	 *
-	 * @since   3.4
-	 */
-	protected function addToolbar()
-	{
-		$canDo = JHelperContent::getActions('com_installer');
-
-		if ($canDo->get('core.edit.state'))
-		{
-			JToolbarHelper::publish('updatesites.publish', 'JTOOLBAR_ENABLE', true);
-			JToolbarHelper::unpublish('updatesites.unpublish', 'JTOOLBAR_DISABLE', true);
-			JToolbarHelper::divider();
-		}
-
-		if ($canDo->get('core.delete'))
-		{
-			JToolbarHelper::deleteList('JGLOBAL_CONFIRM_DELETE', 'updatesites.delete', 'JTOOLBAR_DELETE');
-			JToolbarHelper::divider();
-		}
-
-		if ($canDo->get('core.admin') || $canDo->get('core.options'))
-		{
-			JToolbarHelper::custom('updatesites.rebuild', 'refresh.png', 'refresh_f2.png', 'JTOOLBAR_REBUILD', false);
-		}
-
-		JHtmlSidebar::setAction('index.php?option=com_installer&view=updatesites');
-
-		parent::addToolbar();
-		JToolbarHelper::help('JHELP_EXTENSIONS_EXTENSION_MANAGER_UPDATESITES');
-	}
-}
